add Database
сохранение стримов и сообщений в бд

+++ TinkoffMessenger/app/src/main/java/ru/myacademyhomework/tinkoffmessenger/BottomSheetListener.kt
@@ -1,5 +1,5 @@
 package ru.myacademyhomework.tinkoffmessenger
 
 fun interface BottomSheetListener {
-    fun itemEmojiClicked(emoji: String, idMessage: Int, position: Int)
+    fun itemEmojiClicked(emoji: String, idMessage: Long, position: Int)
 }
