+++ TinkoffMessenger/app/src/main/java/ru/myacademyhomework/tinkoffmessenger/chatFragment/BottomSheetViewHolder.kt
@@ -10,7 +10,7 @@ class BottomSheetViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) 
 
     private val emojiTextView = itemView.findViewById<TextView>(R.id.emoji_bottomsheet)
 
-    fun onBind(emoji: String, idMessage: Int, positionMessage: Int, listener: BottomSheetListener) {
+    fun onBind(emoji: String, idMessage: Long, positionMessage: Int, listener: BottomSheetListener) {
         emojiTextView.text = emoji
         emojiTextView.setOnClickListener {
             listener.itemEmojiClicked(emoji = emoji, idMessage = idMessage, position = positionMessage)
